[#noissue] Split the spring webmvc ITs by version and add a spring 7 IT
- Initialize the spring 7 DispatcherServlet IT with java config
- Split the spring 5 webmvc IT into 5.1 / 5.2 / 5.3 and clean up the split ITs
- Target JDK 17 and spring 6 in the spring-6-it build
