+++ agent-module/plugins-it/spring-it/spring-6-it/pom.xml
@@ -29,10 +29,10 @@
     <packaging>jar</packaging>
 
     <properties>
-        <jdk.version>1.8</jdk.version>
-        <jdk.home>${env.JAVA_8_HOME}</jdk.home>
-        <spring.version>${spring5.version}</spring.version>
-        <jakarta.servlet.version>${jakarta.servlet5.version}</jakarta.servlet.version>
+        <jdk.version>17</jdk.version>
+        <jdk.home>${env.JAVA_17_HOME}</jdk.home>
+        <spring.version>${spring6.version}</spring.version>
+        <jakarta.servlet.version>${jakarta.servlet6.version}</jakarta.servlet.version>
     </properties>
 
     <dependencies>
